+++ src/DocLite/Store/InMemoryStoreAdapter.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace DocLite.Store
 {
@@ -24,11 +25,20 @@ namespace DocLite.Store
             // no op
         }
 
-        public TValue this[TKey index]
+        TKey IStore<TKey, TValue>.this[int index]
         {
-            get { return _dictionary[index]; }
-            set { _dictionary[index] = value; }
+            get { return _dictionary.Keys.ElementAt(index); }
         }
+
+        public TValue this[TKey key]
+        {
+            get { return _dictionary[key]; }
+            set { _dictionary[key] = value; }
+        }
+
+        public ICollection<TKey> Keys { get { return _dictionary.Keys; } }
+        
+        public int Count { get { return _dictionary.Count; } }
 
         public void Dispose()
         {