+++ src/DocLite/Store/PersistentStoreAdapter.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using Microsoft.Isam.Esent.Collections.Generic;
 
 namespace DocLite.Store
@@ -30,11 +31,20 @@ namespace DocLite.Store
             _dictionary.Flush();
         }
 
-        public TValue this[TKey index]
+        TKey IStore<TKey, TValue>.this[int index]
         {
-            get { return _dictionary[index]; }
-            set { _dictionary[index] = value; }
+            get { return _dictionary.Keys.ElementAt(index); }
         }
+
+        public TValue this[TKey key]
+        {
+            get { return _dictionary[key]; }
+            set { _dictionary[key] = value; }
+        }
+
+        public ICollection<TKey> Keys { get { return _dictionary.Keys; } }
+
+        public int Count { get { return _dictionary.Count; } }
 
         public void Dispose()
         {